remove shadowing of `Flow` in src/prefect/flows.py
The previous code uses the `prefect.server.schemas.core.Flow` class within the definition of `prefect.flows.Flow`. This
works because the latter is not yet defined at the point where the annotation is evaluated. However, this is highly
confusion. Not only for human readers, but it also trips up pyright which fails to unambigously locate the definition of
`Flow` in a statement `from prefect.flows import Flow`.

This commit has no runtime effect, but solves the above issues.